Initial version
Though the backend supports per-commit querying, we currently make an
inefficient query to gather all runs.  The backend needs to support
querying by projectid and test name or any arbitrary key/values.  The
frontend needs to be hosted somewhere so that the HTML5 history API will
function properly when implemented.